+++ src/main/java/com/example/financeapp/modules/orderbook/OrderController.java
@@ -1,8 +1,6 @@
 package com.example.financeapp.modules.orderbook;
 
-import com.example.financeapp.modules.account.Account;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.RequestEntity;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -23,9 +21,17 @@ public class OrderController {
 
 
     @PostMapping
-    public ResponseEntity<Order> createOrder(@RequestBody @Valid Order request){
+    public ResponseEntity<String> createOrder(@RequestBody @Valid Order request){
         Order order = orderService.createOrder(request);
-        return ResponseEntity.ok(order);
+        return ResponseEntity.ok("Success");
     }
 
+    @PutMapping
+    public ResponseEntity<String> updateOrder() {
+        return ResponseEntity.ok("Success");
+    }
+
+    @DeleteMapping ResponseEntity<String> deleteOrder() {
+        return ResponseEntity.ok("Success");
+    }
 }
